pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-voiceio"
version = "0.3.0"
description = "Speak → text, locally, instantly."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Hugo Montenegro" }]
keywords = ["voice", "speech-to-text", "whisper", "linux", "windows", "dictation", "wayland", "ibus"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "onnxruntime>=1.16.0",
    "wordfreq>=3.0",
    "evdev>=1.6.0; sys_platform == 'linux'",
    "pynput>=1.7.6; sys_platform == 'win32'",
    "pynput>=1.7.6; sys_platform == 'darwin'",
    "pyperclip>=1.8.0; sys_platform == 'win32'",
    "win11toast>=0.36; sys_platform == 'win32'",
]

[project.optional-dependencies]
x11 = ["pynput>=1.7.6"]
mac = ["pynput>=1.7.6"]
win = ["pynput>=1.7.6", "pyperclip>=1.8.0", "win11toast>=0.36"]
tray = ["pystray>=0.19", "Pillow>=10.0"]
tts = ["piper-tts>=1.2.0"]
tts-cloud = ["edge-tts>=6.1.0"]
dev = ["pytest>=7.0", "pytest-mock"]

[project.urls]
Homepage = "https://github.com/Hugo0/voiceio"
Repository = "https://github.com/Hugo0/voiceio"
Issues = "https://github.com/Hugo0/voiceio/issues"
Changelog = "https://github.com/Hugo0/voiceio/releases"

[project.scripts]
voiceio = "voiceio.cli:main"
# Legacy aliases (prefer: voiceio toggle/doctor/setup/test)
voiceio-toggle = "voiceio.cli:_cmd_toggle"
voiceio-doctor = "voiceio.cli:_cmd_doctor_legacy"
voiceio-setup = "voiceio.cli:_cmd_setup"
voiceio-test = "voiceio.cli:_cmd_test"

[tool.ruff]
lint.per-file-ignores."voiceio/ibus/engine.py" = ["E402"]

[tool.setuptools.packages.find]
include = ["voiceio*"]

[tool.setuptools.package-data]
"voiceio.sounds" = ["*.wav"]
"voiceio.models" = ["*.onnx"]
